+++ Assets/Resources/Scripts/Combat/SkillInitGold.cs
@@ -13,7 +13,7 @@ public class SkillInitGold : Skill
         EffectManager.PlaySkillEffect(owner, skillCfg.HitEffect);
 
         var goldAdd = Math.Max(owner.GetAttr(skillCfg.Attr) * skillCfg.SkillAttrRate, skillCfg.StrengthInt);
-        owner.GetPlayerInfo().AddGold(goldAdd);
+        owner.GetPlayerInfo().AddGold((int)goldAdd);
 
         WorldManager.Instance.AddBattleText(goldAdd.ToString() + "金", owner.transform.position, new UnityEngine.Vector2(0, 60), Color.yellow, 2);
     }
